HtmlBridge Phase 3 (P3.30): extract document.createEvent into LegacyEventBinding
document.createEvent (the legacy DOM Events Level 3 factory) is now the co-located
LegacyEventBinding module (namespace Broiler.HtmlBridge.Dom.Features); was the bridge's
~320-line JsRegistrationCreateEvent033Core.

The builder returns a plain event object pre-populated with the union of
UI/Mouse/Keyboard/Wheel/Custom event fields plus the legacy init*Event and
propagation-control methods. It builds a self-contained JS object with closures over its
own state and touches no bridge instance state, so — like ConsoleBinding / CryptoBinding
— it is a pure static class with no host contract (registered by method group).

This is the single biggest callback moved out of the grab-bag:
JsFunctionCallbacks/Registration.cs 653 -> 328 lines.

Behaviour-preserving; no public-API change (module internal). Verified:
Broiler.HtmlBridge.Dom builds clean; new LegacyEventBindingModuleTests (co-location +
callback-moved-off-bridge guards + a characterization: createEvent('Event'), initEvent
sets type/bubbles/cancelable, preventDefault on a cancelable event sets defaultPrevented,
propagation methods present) pass alongside the DomEvents (51, a heavy createEvent user),
DomEventsEdgeCase and architecture-guard suites.
