+++ src/Broiler.HtmlBridge.Dom/DomBridge/Registration/Document.cs
@@ -46,8 +46,8 @@ public sealed partial class DomBridge
         document.FastAddValue((KeyString)"createAttribute", new JSFunction((in a) => Dom.Features.DocumentFactoryBinding.CreateAttribute(this, context, in a), "createAttribute", 1), JSPropertyAttributes.EnumerableConfigurableValue);
         document.FastAddValue((KeyString)"createDocumentFragment", new JSFunction((in a) => Dom.Features.DocumentFactoryBinding.CreateDocumentFragment(this, in a), "createDocumentFragment", 0), JSPropertyAttributes.EnumerableConfigurableValue);
 
-        // document.createEvent(type) — DOM Events Level 3
-        document.FastAddValue((KeyString)"createEvent", new JSFunction(JsRegistrationCreateEvent033Core, "createEvent", 1), JSPropertyAttributes.EnumerableConfigurableValue);
+        // document.createEvent(type) — DOM Events Level 3 (Phase 3: co-located LegacyEventBinding module)
+        document.FastAddValue((KeyString)"createEvent", new JSFunction(Dom.Features.LegacyEventBinding.Create, "createEvent", 1), JSPropertyAttributes.EnumerableConfigurableValue);
     }
 
     private void RegisterDocumentWriting(JSObject document)
